Fail more gracefully and do not block build
When there is no connection or wrong data build is terminated. This
seems like a wrong behavior as the plugin actually blocks any sphinx
build if its blocks cannot be resolved. To fix this the whole try-except
block is wrapped around connection items, which makes build just display
the error, but still continue.

Tested with setting PC offline and building example documentation.