Provide error text outside of the table
There was an issue with Latex rendering when we reported the error, but
did not complete table row with empty number of remaining columns.
Current solution closes the table (header displayed only) and provides
the normal paragraph of text below the table.

html renderer had no problems with previous implementation, but it is
nice to have a decent pdf rendering.