Login to server once, retrieve defects with each node
We were logging in and retrieving stream and project information from
coverity server with each node. This is not needed and is now done when
builder initializes. It also enabled us to print nicer warning outputs
when there was a login error, but we still do not fail build with it. It
is up for external tools (mlx.warnings) to error in case warnings are
unwanted.